@@ -1,49 +1,41 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, AfterViewInit } from '@angular/core';
 import { Router } from '@angular/router';
 
 import { AuthService } from './services/auth.service';
 import { User } from './models/user';
 
+declare var M: any;
+
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.css'],
-  providers: [AuthService]
 })
-export class AppComponent implements OnInit {
+export class AppComponent implements OnInit, AfterViewInit {
   public constructor(
-    private authService: AuthService,
-    private router: Router
+    public authService: AuthService,
+    public router: Router
   ) {}
 
   ngOnInit() {
-    this.authService.syncLoginStatus();
+    //this.authService.syncCurrentUser();
   }
 
-  private logout() {
+  ngAfterViewInit() {
+    let elems = document.querySelectorAll('.sidenav');
+    let instances = M.Sidenav.init(elems);
+  }
+
+  logout() {
     this.authService.logout();
     this.router.navigate(['/logged-out']);
   }
 
-  private loggedIn(): boolean {
-    return this.authService.getUser() !== undefined;
-  }
-
-  private getUserRealName(): String {
-    return this.authService.getUser().realName;
-  }
-
-  private isStudent(): boolean {
-    return this.authService.getUser().role === 'ROLE_STUDENT';
-  }
-
-  private isTeacher(): boolean {
-    return this.authService.getUser().role === 'ROLE_TEACHER';
-  }
-
-  private isAdmin(): boolean {
-    return this.authService.getUser().role === 'ROLE_ADMIN';
-  }
+  public isLinkActive(url) {
+    let charPos = this.router.url.indexOf('?');
+    let cleanUrl = charPos !== -1 ? this.router.url.slice(0, charPos) : this.router.url;
+    return (cleanUrl === url);
+}
 
   public title: string = 'e-Napló';
 }
