@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, AfterViewInit } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { DatePipe } from '@angular/common';
 
@@ -9,18 +9,20 @@ import { Mark } from '../../models/mark';
 import { AuthService } from '../../services/auth.service';
 import { UserService } from '../../services/user.service';
 
+declare var M: any;
+
 @Component({
   selector: 'app-student-subject-list-view',
   templateUrl: './student-subject-list-view.component.html',
   styleUrls: ['./student-subject-list-view.component.css'],
   providers: [UserService, DatePipe]
 })
-export class StudentSubjectListViewComponent implements OnInit {
+export class StudentSubjectListViewComponent implements OnInit, AfterViewInit {
   private user: User;
   private users: User[];
-  private subjects: Subject[];
-  private marks: Mark[];
-  private average: Number;
+  subjects: Subject[];
+  marks: Mark[];
+  average: Number;
 
   constructor(
     private route: ActivatedRoute,
@@ -33,30 +35,38 @@ export class StudentSubjectListViewComponent implements OnInit {
   ) {
   }
 
-  ngOnInit() {
+  async ngOnInit() {
     this.setUser();
-    this.update();
+    await this.update();
+
+    console.log('Userrrrr: ' + this.authService.getUser())
+  }
+
+  ngAfterViewInit() {
+
   }
 
   private setUser(): void {
     this.user = this.authService.getUser();
   }
 
-  private update(): void {
-    const userId = this.user.id;
-    // Tárgyak
-    this.userService.getSubjectsByUser(userId).subscribe((subjects) => {
-      this.subjects = subjects as Subject[];
-      this.calculateAverages();
-    });
-    // Felhasználók
-    this.userService.getUsers().subscribe((users) => {
-      this.users = users as User[];
-    });
+  private initMaterializeCollapsible() {
+    let elems = document.querySelectorAll('.collapsible');
+    M.Collapsible.init(elems);
+  }
+
+  async update() {
+    this.users = await this.userService.findAll();
+    this.subjects = await this.userService.getSubjects(this.user.id);
+    this.calculateAverages();
+
+    setTimeout(() => {
+      this.initMaterializeCollapsible();
+    }, 500)
   }
 
   // Kiszámolja, hogy az egyes tárgyakhoz tartozó jegyek átlaga mennyi (jelen esetben ez egy felhasználóra történik meg)
-  private calculateAverages(): void {
+  calculateAverages(): void {
     if (this.subjects !== undefined) {
       for (let i = 0; i < this.subjects.length; i++) {
         let marks = this.subjects[i].marks;
@@ -67,7 +77,7 @@ export class StudentSubjectListViewComponent implements OnInit {
   }
 
   // A már tárgyanként kiszámolt átlagokat átlagolja
-  private getTotalAverage(): Number {
+  getTotalAverage(): Number {
     if (this.subjects !== undefined) {
       let sum = this.subjects.reduce((a, b) => a + b.average, 0);
       return parseFloat((sum/this.subjects.length).toFixed(2));
@@ -75,11 +85,11 @@ export class StudentSubjectListViewComponent implements OnInit {
     return 0;
   }
 
-  private transformTimestamp(timestamp: number): string {
+  transformTimestamp(timestamp: number): string {
     return this.datePipe.transform(timestamp, 'y. MM. dd. HH:mm');
   }
 
-  private getUserName(userId: number): String {
+  getUserName(userId: number): String {
     let filtered = this.users.filter((user) => user.id === userId);
     return filtered.length == 1 ? filtered[0].realName : "Ismeretlen felhasználó";
   }
