@@ -1,23 +1,20 @@
+// Angular
 import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs';
 
-import { Subject } from '../models/subject';
-import { Group } from '../models/group';
-import { User } from '../models/user';
+// App
+import { Subject } from '@app/models/subject';
+import { HttpService } from '@app/services/http.service';
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class SubjectService {
-  private static api: string = 'http://localhost:8080/api/subjects';
-  private handleError;
-  
-  constructor(
-    private http: HttpClient
-  ) { }
+  private baseUrl = 'http://localhost:8080/api/subjects';
 
-  public getSubjects(): Observable<Subject[]> {
-    return this.http.get(SubjectService.api) as Observable<Subject[]>;
+  constructor(private httpService: HttpService) { }
+
+  public getSubjects(): Promise<Subject[]> {
+    return this.httpService.get(`${this.baseUrl}`);
   }
 
-  // További funkciók
 }
